[flang][OpenMP] Move TargetOMPContext to shared FlangOMPContext (NFC)
Moving the class to shared code makes it available for reuse by
forthcoming DECLARE VARIANT lowering without any functional change to
existing metadirective lowering.
